Module Asai_lsp.Make

This module provides a rudimentary and incomplete implementation of the LSP protocol.

Note: many features are missing and it does not handle positionEncoding.

Parameters

module Message : Asai.MinimumSigs.Message

Signature

val start : source:string option -> init:(root:string option -> unit) -> load_file:(display:(Message.t Asai.Diagnostic.t -> unit) -> string -> unit) -> unit

run ~init ~load_file starts the LSP server with the two callbacks init and load_file.

parameter source
The source of a LSP diagnostic, that is, a "human-readable string describing the source of this diagnostic."

parameter init
The callback to initiate the loading of a workspace. The root parameter is the workspace

parameter load_file
The callback to load the file.